materializations: use InfoSchema to do `Apply` via `Applier` for dynamodb and elasticsearch
